F-02 Remove order page container. Refactoring map with useAppSelector and useAppDispatch
