+++ src/App.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom'
 
 import MainPage from './routes/mainPage';
-import OrderPageContainer from './containers/orderPageContainer';
+import OrderPage from './routes/orderPage';
 
 export default function App() {
   return (
     <Switch>
       <Route exact path='/' component={MainPage} />
-      <Route exact path='/order' component={OrderPageContainer} />
+      <Route exact path='/order' component={OrderPage} />
     </Switch>
   );
 }
