+++ src/routes/orderPage/index.tsx
@@ -3,7 +3,6 @@ import { Tabs } from 'antd';
 const { TabPane } = Tabs;
 
 import './styles.scss'
-import { IOrderPageContainer } from '../../containers/orderPageContainer';
 import useToggle from '../../hooks/useToggle'
 import Aside from '../../components/aside'
 import HamburgerMenu from '../../components/hamburgerMenu'
@@ -13,7 +12,7 @@ import StepOne from '../../components/steps/stepOne';
 import StepTwo from '../../components/steps/stepTwo';
 import StepThree from '../../components/steps/stepThree';
 
-export default function OrderPage(props: IOrderPageContainer) {
+export default function OrderPage() {
   const [isOpen, setIsOpen] = useToggle(false)
 
   return (
@@ -29,16 +28,7 @@ export default function OrderPage(props: IOrderPageContainer) {
 
         <Tabs type="card" className='order-page__tabs' >
           <TabPane tab="Местоположение" key="1" >
-            <StepOne
-              town={props.town}
-              pickUp={props.pickUp}
-              coordinatesByPickedTown={props.coordinatesByPickedTown}
-              markers={props.markers}
-              onAddMark={props.handleAddMarker}
-              onSetTown={props.handleSelectTown}
-              onSetPickUp={props.handleSelectPickUp}
-              handleSelectCoordinates={props.handleSelectCoordinates}
-            />
+            <StepOne />
           </TabPane>
 
           <TabPane tab="Модель" key="2" >
