+++ src/routes/orderPage/styles.scss
@@ -67,7 +67,7 @@
         height: 0;
         border-style: solid;
         border-width: 4px 0 4px 6px;
-        border-color: transparent transparent transparent #999999; 
+        border-color: transparent transparent transparent $grey-strong-color; 
       }
     }
   }
